ShippingService/requirements.txt
prometheus-client
gunicorn
whitenoise
Django>=4.2
djangorestframework>=3.15
drf-yasg>=1.21.4
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31.0
python-json-logger>=2.0.2
django-prometheus>=2.3.1
django-extensions>=3.2.3
